Extract inventory worker into hook
- Move inventory WebWorker logic into a new useInventoryWorker hook and refactor App to use it. Introduces src/hooks/useInventoryWorker.ts which encapsulates worker lifecycle, message handling, and inventory cache updates; App.tsx now imports and calls useInventoryWorker and replaces inline worker setup. Also: make fetchAccounts and resetAccountLoading useCallback, tidy imports (remove appendUniqueItems from App), and simplify Topbar form defaultValues to read directly from the store. Update appStore to track cartItemIds (Set) for O(1) membership checks and adjust addToCart/removeFromCart/clearCart accordingly; remove the deprecated AppActions type and an unused store subscription. These changes encapsulate worker behavior, reduce App complexity, and improve cart performance.
