+++ src/stores/appStore.ts
@@ -40,27 +40,6 @@ interface AppState {
   loginOpen: boolean;
 }
 
-interface AppActions {
-  setRealm: (realm: string) => void;
-  setGameType: (gameType: string) => void;
-  setGameMode: (gameMode: string) => void;
-  setGameClass: (gameClass: string) => void;
-  setApiUrl: (apiUrl: string) => void;
-  setUsername: (username: string) => void;
-  setGameName: (gameName: string) => void;
-  setCart: (cart: InventoryItem[]) => void;
-  addToCart: (item: InventoryItem) => void;
-  removeFromCart: (item: InventoryItem) => void;
-  clearCart: () => void;
-  setPassword: (password: string) => void;
-  setCartOpen: (cartOpen: boolean) => void;
-  setRecentDropsOpen: (open: boolean) => void;
-  setInventory: (inventory: InventoryItem[]) => void;
-  setLoadingInventory: (loading: boolean) => void;
-}
-
-export type AppStore = AppState & AppActions;
-
 export type AccountDataCache = {
   realm: string;
   accountName: string;
@@ -174,21 +153,25 @@ export const setCart = (cart: InventoryItem[]) => {
 
 export const addToCart = (item: InventoryItem) => {
   useAppStore.setState((state) => {
-    if (state.cart.some((i) => i.itemid === item.itemid)) {
-      return { cart: state.cart };
-    }
-    return { cart: [...state.cart, item] };
+    if (state.cartItemIds.has(item.itemid)) return {};
+    const cart = [...state.cart, item];
+    const cartItemIds = new Set(state.cartItemIds);
+    cartItemIds.add(item.itemid);
+    return { cart, cartItemIds };
   });
 };
 
 export const removeFromCart = (item: InventoryItem) => {
   useAppStore.setState((state) => {
-    return { cart: state.cart.filter((i) => i.itemid !== item.itemid) };
+    const cart = state.cart.filter((i) => i.itemid !== item.itemid);
+    const cartItemIds = new Set(state.cartItemIds);
+    cartItemIds.delete(item.itemid);
+    return { cart, cartItemIds };
   });
 };
 
 export const clearCart = () => {
-  useAppStore.setState({ cart: [] });
+  useAppStore.setState({ cart: [], cartItemIds: new Set() });
 };
 
 export const setCartOpen = (cartOpen: boolean) => {
@@ -339,17 +322,3 @@ export const updateCachedDrops = () => {
       });
   }
 };
-
-useAppStore.subscribe(
-  (state) => state.cart,
-  (cart) => {
-    const newCartItemIds = new Set(cart.map((i) => i.itemid));
-    const currentCartItemIds = useAppStore.getState().cartItemIds;
-    if (
-      newCartItemIds.size !== currentCartItemIds.size ||
-      [...newCartItemIds].some((id) => !currentCartItemIds.has(id))
-    ) {
-      useAppStore.setState({ cartItemIds: newCartItemIds });
-    }
-  },
-);
